# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ExecuTorch Bencher API",
    "description": "Internal Android LLM benchmark collection and analysis service. Provides HTTP endpoints to ingest, store, and retrieve benchmark run results, content-addressed artifacts, and registered model assets.",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List registered model assets, optionally narrowed to one content hash.\nNot paginated - matches the current implementation's real, unpaginated\nresponse shape. See `specs/ingestion-service` - \"Service exposes model\nlookup by content hash\".",
        "operationId": "listModels",
        "parameters": [
          {
            "name": "sha256",
            "in": "query",
            "description": "Return only the model asset with this content hash (lowercase\n64-character hexadecimal SHA-256), so a collector that knows a\nmodel file's hash can find its asset ID. An empty list when no\nasset matches.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The matching model assets (all of them without a filter).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelAssetResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "`sha256` is not a well-formed digest.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal storage error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/register": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Register an external `.pte` model file by path, without copying it.\nRegistering a file whose SHA-256 already matches a registered model\nreuses the cached checksum and existing record rather than rehashing or\nduplicating it.",
        "operationId": "registerModel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The model was registered, or an existing record with the same SHA-256 was reused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAssetResponse"
                }
              }
            }
          },
          "400": {
            "description": "The path does not exist or is not a regular file.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal storage error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{id}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Fetch a single registered model asset by ID.",
        "operationId": "getModel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The model asset's identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The model asset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAssetResponse"
                }
              }
            }
          },
          "404": {
            "description": "No model asset exists with this ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{id}/verify": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Trigger explicit full verification of a model's checksum: recalculates\nits SHA-256 from the current file content, updates its last-verified\ntimestamp, and reports whether it still matches the registered value.\nOnly implemented for external-mode models.",
        "operationId": "verifyModel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The model asset's identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification completed; the asset's cached checksum, availability, and last-verified timestamp were refreshed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAssetResponse"
                }
              }
            }
          },
          "404": {
            "description": "No model asset exists with this ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Verification for managed-mode models is not yet implemented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "description": "The `error` object inside an [`ApiErrorResponse`].",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable, machine-readable failure category. Current values: `invalid_request`,\n`not_found`, `artifact_file_missing`, `payload_too_large`, `conflict`,\n`not_implemented`, `internal_error`. This field is an open string, not a\nclosed enum - new categories may be added as the API grows.",
            "example": "invalid_request"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the failure. Not guaranteed stable across\nversions or releases - match on `code`, not this text.",
            "example": "gpu_clock_mhz must be greater than zero"
          },
          "details": {
            "description": "Optional structured context about the failure, e.g. which field was invalid."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque identifier correlating this error with server-side logs, when available."
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "The consistent JSON envelope every rejected or failed request responds with.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        }
      },
...
      "ModelAssetResponse": {
        "type": "object",
        "description": "A registered model asset's metadata.",
        "required": [
          "id",
          "sha256",
          "original_name",
          "size_bytes",
          "model_format",
          "storage_mode",
          "registered_at",
          "available"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sha256": {
            "type": "string",
            "description": "Lowercase 64-character hexadecimal SHA-256 digest of the model file.",
            "example": "b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944"
          },
          "original_name": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "model_format": {
            "type": "string"
          },
          "storage_mode": {
            "$ref": "#/components/schemas/ModelStorageMode"
          },
          "external_path": {
            "type": [
              "string",
              "null"
            ],
            "description": "The server-side filesystem path this external model was registered\nfrom. `null` for managed-mode models. Currently returned to every\ncaller as-is - no authentication exists in this service to restrict\nit to authorized clients."
          },
          "registered_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Absent until the model has been explicitly verified at least once."
          },
          "available": {
            "type": "boolean",
            "description": "Whether the registered file currently resolves on disk."
          }
        }
      },
...
      "ModelStorageMode": {
        "type": "string",
        "enum": [
          "external",
          "managed"
        ]
      },
...
      "RegisterModelRequest": {
        "type": "object",
        "required": [
          "path"
        ],
        "properties": {
          "path": {
            "type": "string",
            "description": "Absolute path to an existing `.pte` file on the server's filesystem.\nExternal mode never copies this file - see\n`specs/artifact-storage` - \"External model assets are registered\nonce without copying\". Registering the same file's SHA-256 again\nreuses the cached checksum rather than rehashing or copying it.",
            "example": "/data/models/llama-3-8b-instruct.pte"
          }
        }
      },
...
    }
  },
...
}